Converted assistant commands to json format, added custom logging function, moved authenticate to helper module
