--- authenticate.py
@@ -1,9 +1,0 @@
-key_set = set()
-secret_f = open('secret')
-line = secret_f.readline()
-while line:
-    key_set.add(line)
-    line = secret_f.readline()
-
-def is_valid_key(key):
-    return key and key in key_set
